refactor: unify starter catalogs into one getStarters source
Collapse officialTemplates.ts and DefaultTemplatesService.ts into a single
src/services/starters.ts with one Starter type, one getStarters() factory,
and a surface discriminator. TemplatesController serves /api/templates/official
and /api/templates/defaults as filtered views over the one list, stripping the
internal surface flag so both endpoints return byte-identical output to main.

A characterization test snapshots the exact pre-refactor output of both
factories (volatile noteType.id excluded) and asserts the unified module
reproduces it field-for-field. The #2328 field-resolves-against-flds validator
now runs against every starter in the unified list. The two service test
suites are merged into starters.test.ts.

No user-visible change.

Fixes #2333
